review comments: moved functions to utils
